+++ Chartboost/CHANGELOG.md
@@ -1,4 +1,12 @@
 ## Changelog
+  * 7.3.1.1
+    * **Note**: This version is only compatible with the 5.5.0+ release of the MoPub SDK.
+    * Add the `ChartboostAdapterConfiguration` class to: 
+         * pre-initialize the Chartboost SDK during MoPub SDK initialization process
+         * store adapter and SDK versions for logging purpose
+    * Streamline adapter logs via `MoPubLog` to make debugging more efficient. For more details, check the [Android Initialization guide](https://developers.mopub.com/docs/android/initialization/) and [Writing Custom Events guide](https://developers.mopub.com/docs/android/custom-events/).
+    * Allow supported mediated networks and publishers to opt-in to process a user’s personal data based on legitimate interest basis. More details [here](https://developers.mopub.com/docs/publisher/gdpr-guide/#legitimate-interest-support).
+
   * 7.3.1.0
     * This version of the adapters has been certified with Chartboost 7.3.1.